+++ NEWS.md
@@ -1,3 +1,17 @@
+Version 0.9.4 (2023.10.31)
+--------------------------
+
+* Switched to `getrandom` in library (demo uses `rand` as before)
+
+* Replaced `zmq_msg_init_data()` with `zmq_msg_init_size()`
+
+* `ZMQ_BLOCKY` is set to 0 at once for entire context, instead of `ZMQ_LINGER` for each socket
+
+* Tiny changes and fixes
+
+* Added links to Go version
+
+
 Version 0.9.3 (2023.10.12)
 --------------------------